requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
attrs==26.1.0
croniter==6.2.2
discord-py-interactions==5.16.0
discord-typings==0.9.0
emoji==2.15.0
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
tomli==2.4.1
typing_extensions==4.15.0
yarl==1.23.0
